parametrized installation build (genomics yes or no, compile orange, compile crs)

+++ install-scripts/mac/macinstallation.sh
@@ -1,29 +1,37 @@
+#!/bin/bash
 ##cvs -d :pserver:tomazc@estelle.fri.uni-lj.si:/d//cvs login
 
+CORNG=0
+CCRS=0
+GEN=1
 
 ## check out orange source and compile orange
 mkdir compiledOrange
-##rm -Rf source
-##cvs -d :pserver:tomazc@estelle.fri.uni-lj.si:/d//cvs export -r HEAD source
-##cd source
-##python makedep.py
-##make -f Makefile.mac
-##cd ..
+if [ $CORNG == 1 ]; then
+  rm -Rf source
+  cvs -d :pserver:tomazc@estelle.fri.uni-lj.si:/d//cvs export -r HEAD source
+  cd source
+  python makedep.py
+  make -f Makefile.mac
+  cd ..
+fi
 mv *.so compiledOrange
 
 
 ## compile orngCRS
-##rm -Rf orngExtn-1_8_1_py23
-##tar -xvzf orngExtn-1_8_1_py23.mac.tar.gz
-##cd orngExtn-1_8_1_py23
-##python setup.py build
-##mv build/lib.darwin-6.8-Power_Macintosh-2.3/_orngCRS.so ../compiledOrange
-##cd ..
-
+if [ $CCRS == 1 ]; then
+  rm -Rf orngExtn-1_8_1_py23
+  tar -xvzf orngExtn-1_8_1_py23.mac.tar.gz
+  cd orngExtn-1_8_1_py23
+  python setup.py build
+  mv build/lib.darwin-6.8-Power_Macintosh-2.3/_orngCRS.so ../compiledOrange
+  cd ..
+fi
 
 ## check out orange modules
 rm -Rf orange doc
 cvs -d :pserver:tomazc@estelle.fri.uni-lj.si:/d//cvs export -r HEAD orange
+cvs -d :pserver:tomazc@estelle.fri.uni-lj.si:/d//cvs export -r HEAD -d orange/OrangeWidgets/Genomics Genomics
 mv orange/doc Orange\ Doc
 
 # remove files we don't want in the installation
@@ -34,11 +42,13 @@ rm orange/OrangeWidgets/Data/OWExampleBuilder.py
 rm orange/OrangeWidgets/Data/OWSubsetGenerator.py
 rm orange/OrangeWidgets/OWLin_Results.py
 rm orange/OrangeWidgets/Other/OWITree.py
-rm -Rf orange/OrangeWidgets/Genomics
+if [ $GEN == 0 ]; then
+  rm -Rf orange/OrangeWidgets/Genomics
+fi
 
 ## after compiling orange, move it out of the path and into the orange directory
 cp compiledOrange/*.so orange
-setenv PYTHONPATH .
+PYTHONPATH=.
 rm -Rf build
 python makeapplication.py --resource=orange build
 tar -C ~/Desktop -xvzf emptyDiskImages/Orange.tar.gz
